Fixed broken projectile spawning.
This commit fixes an existing bug with enemy projectile generation. Previously, instances of the Enemy class
could not properly generate new projectiles aside from the inital group made when the static projectile group
was instantiated. This was because the newest iteration of the Projectile class constructor requires an argument
to determine which graphic it should use and the default use of recycle() can't handle a constructor with non-
optional arguments.

The factory function introduced in this commit binds the proj parameter passed into the fire projectile function
so that it can instantiate a projectile properly when a new one needs to be made. Now enemies can make new
projectiles beyond their initial pool.