Updated PhysicalEnemy size and health variables.
The PhysicalEnemy's hitbox was previously much larger than its sprite, which is now fixed.
Also reduced its health, as the next commit will allow more than a single physicalEnemy to
be in the playState at once. Also made sure that enemies started in the proper state when
init() is called.