Replaced ProjectileSpawner with EnemySpawner.
Since the child classes of Enemy will handle what ProjectileSpawner used to do, ProjectileSpawner
was replaced with a class that spawns enemies. Much of the logic is the same, but enemies are
created and added to the state instead of projectiles.

Additional changes will be needed to handle spawning different types of enemies, but that will
be handled after other enemy types have been set up in the game.