Created an alternate movement function for Enemy
I wanted to make the physical enemy constantly move downward in addition to moving side-to-side,
but the angle-based movement system that Enemy currently used would make it a bit more
complicated than it would need to be. The new componentMovement function solves this problem
by simplifying any movement that would be better expressed with vertical & horizontal components.

The two movement styles can also be combined if a special movement effect is desired. For
example, if we want an enemy that moves in a wave pattern while moving vertically, use the
angledMovement function to move in a horizontal wave, then add in a vertical component
with component movement.