Changed Enemy display group to _grpActors
Previously, enemies were added to _grpEnemies for both logic and display
purposes. These changes make _grpEnemies a purely logical grouping, used
when checking for collisions. This helps it more closely match _grpCharacters,
which is a purely logical grouping for the player controlled actors.

_grpActors is now the display group for both characters and enemies, which
is appropriate. This ensures that all actors moving around the game world will
be properly organized into the foreground/background according to their y-value,
which is more important now that enemies are more likely to overlap the player
characters.