Added collision logic between actors
Enemies and player-controlled characters now have functions to handle collisions
between them. If an enemy runs into the player's energy shield, the capacity will
go up at a steady rate. If an enemy runs into a player character, then both parties
will take damage.

At the moment the damage dealt is rather unbalanced--the mechanic character
is defeated if it runs into an enemy once--but that can be fixed later when all of the
components are in place and working properly.