Set maximum damage MechanicChar can take from attacks
Previously, the mechanic character could be killed after a single
collision with an enemy character. Rather than rework multiple files
to increase the mechanic's survivability, I set a cap on how much
damage the mechanic can take from a single instance of damage.

At the moment, that cap is one, but it can be easily changed in the future
if necessary.