Completed basic UI
This commit adds healthbars for the two player characters on either side of the shield capacity bar.
The UI class was also changed to not inherit from FlxSprite, since it didn't need any of its functions.

Overall, it works reasonably well, although the code in UI could definitely be rewritten to be more
elegant.