Implemented transition between pregame and gameplay
This commit adds a function to UI that fades out and removes the pregame overlay.

PlayState now adds the pregame overlay during normal setup and waits to add
an instance of EnemySpawner to the game. After the player presses a normal
gameplay input, the UI class's new function is called to fade out the overlay and
an instance of EnemySpawner is added to the game.