Implemented dead state for RobotChar
This commit changes the RobotChar class to die properly. Before this commit,
The robot would de-activate its shield and call kill() when it ran out of health, which
caused it to disappear from the screen. Now the robot becomes broken,
unresponsive to player input, and moves down until it leaves the screen. It can still
collide wth the MechanicChar normally, but it otherwise has no collision.

This behavior upon death is definitely preferable, because the game can continue
after the robot dies. It wouldn't make sense to leave the robot still against a moving
background, so the robot gets left behind by the always-running mechanic.