Fixed a game-crashing bug involving projectile pools
Since projectile pools are a static variable of their enemy class,
they need to be set to null before the FlxG can restart the state.
If they aren't set to null (and subsequently replaced after restart)
then an invalid array access exception occurs.

I'm not totallly sure about the technical reasons behind this, but
using this function to set the projectile pools variables to null just
before the reset occurs fixes it.