Implemented Robot and Shield animations
Was relatively straightforward, except for the shield. Decided to split the
graphic and the logical components of the shield up because it seemed
like it would be easiest that way.  As a result, EnergyShield now inherits from
FlxObject instead of FlxSprite, and RobotChar is now responsible for setting up
an FlxSprite that contains the shield graphic and animation information.

A couple of hacky things were done to make the shield graphic appear and
disappear at the right times, which is a bit unfortunate. Shouldn't be too hard
to figure out how it works, though. To make the shield be behind other graphics
I put it in its own flxGroup in the playstate, _grpBackgroundSprites. On the whole,
it works quite well, though.