Fixed a bug involving inconsistent robot coloring after death
Previously, the robot always did its brokenTransition upon death. If the robot was already broken,
however, it would change the robot's coloring back to normal, due to how the adjustColor function
works. This makes sure that the robot only does the brokenTransition if it isn't already broken,
and it ensures that the robot can't die twice (once from damage and once when the mechanic dies).